refactor(app-frame): split app_frame.py handler into topic mixins (#802)
Decompose the 452-LOC residual catch-all handler mixin into three focused
topic mixins plus a smaller residual, following the package's established
topic-split-mixin convention (composed into AppFrame over AppFrameProto).

- toolbar_menu.py (151): ToolbarMenuHandlers — _open_toolbar_settings_menu,
  the _append_menu_item/_append_radio_submenu builders, and the
  _apply_deck_source/_apply_language/_apply_average_method/_apply_average_hours
  setters.
- window_layout.py (144): WindowLayoutHandlers — _save_window_settings,
  _apply_window_preferences, _target_display_client_area, _apply_min_size, and
  the collapse toggles (kept together since _apply_window_preferences calls
  _set_*_collapsed and _relayout_after_toggle calls _apply_min_size).
- deck_render.py (100): DeckRenderHandlers — fetch_archetypes,
  _clear_deck_display, _render_current_deck/_render_pending_deck,
  _populate_archetype_list, _on_deck_download_success, _update_stats,
  _handle_image_downloaded.
- app_frame.py (125, residual): AppFrameHandlersMixin — session restore, aux
  openers, and both debounce timers (kept together: they share the single
  EVT_TIMER dispatch surface).

Behavior-preserving: methods moved verbatim, all instance state still lives on
AppFrame and is reached via self/AppFrameProto. New mixins registered in
handlers/__init__.py and AppFrame's MRO (wx.Frame stays last).
