+++ widgets/frames/app_frame/handlers/__init__.py
@@ -5,6 +5,7 @@ from widgets.frames.app_frame.handlers.card_tables import CardTablesHandler
 from widgets.frames.app_frame.handlers.child_windows import ChildWindowHandlers
 from widgets.frames.app_frame.handlers.data_loading import DataLoadingHandlers
 from widgets.frames.app_frame.handlers.deck_content import DeckContentHandlers
+from widgets.frames.app_frame.handlers.deck_render import DeckRenderHandlers
 from widgets.frames.app_frame.handlers.research import DeckResearchHandlers
 from widgets.frames.app_frame.handlers.sideboard_guide_entries import (
     SideboardGuideEntryHandlers,
@@ -18,6 +19,8 @@ from widgets.frames.app_frame.handlers.sideboard_guide_persistence import (
 from widgets.frames.app_frame.handlers.sideboard_guide_record import (
     SideboardGuideRecordHandlers,
 )
+from widgets.frames.app_frame.handlers.toolbar_menu import ToolbarMenuHandlers
+from widgets.frames.app_frame.handlers.window_layout import WindowLayoutHandlers
 
 __all__ = [
     "AppFrameHandlersMixin",
@@ -25,9 +28,12 @@ __all__ = [
     "ChildWindowHandlers",
     "DataLoadingHandlers",
     "DeckContentHandlers",
+    "DeckRenderHandlers",
     "DeckResearchHandlers",
     "SideboardGuideEntryHandlers",
     "SideboardGuideImportExportHandlers",
     "SideboardGuidePersistenceHandlers",
     "SideboardGuideRecordHandlers",
+    "ToolbarMenuHandlers",
+    "WindowLayoutHandlers",
 ]
